Improve tests for Add a buyer domain
There aren't a lot of tests for our admin index page, but at least they have their own file now.

Changed test to assert_flashes as well as test for the desired success message on the destination page.
