+++ tests/app/main/views/test_services.py
@@ -19,32 +19,6 @@ from ...helpers import LoggedInApplicationTest
 from dmutils import s3
 
 
-class TestIndex(LoggedInApplicationTest):
-    @mock.patch('app.main.views.services.data_api_client')
-    def test_index_shows_frameworks_in_standstill_or_live(self, data_api_client):
-        data_api_client.find_frameworks.return_value = {'frameworks': [
-            {'id': 1, 'frameworkAgreementVersion': None, 'name': 'Framework 1', 'slug': 'framework-1',
-             'status': 'standstill'},
-            {'id': 2, 'frameworkAgreementVersion': 'v1.0', 'name': 'Framework 2', 'slug': 'framework-2',
-             'status': 'live'},
-            {'id': 3, 'frameworkAgreementVersion': None, 'name': 'Framework 3', 'slug': 'framework-3',
-             'status': 'open'},
-        ]}
-
-        response = self.client.get('/admin')
-        data = response.get_data(as_text=True)
-
-        assert 'Download Framework 1 agreements' in data
-        assert 'Approve Framework 2 agreements for countersigning' in data
-
-        # Agreements should be in reverse-chronological order.
-        assert data.index('Approve Framework 2 agreements for countersigning') < \
-            data.index('Download Framework 1 agreements')
-
-        # Only standstill/live agreements should be listed.
-        assert 'Download Framework 3 agreements' not in data
-
-
 class TestServiceFind(LoggedInApplicationTest):
 
     def test_service_find_redirects_to_view_for_valid_service_id(self):
